+++ tests/es_api_spec/test_spec_builder.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 from peek import __file__ as package_root
-from peek.js_spec.spec_builder import JsSpecParser, JsSpecEvaluator
+from peek.es_api_spec.spec_builder import JsSpecParser, JsSpecEvaluator
 
 package_root = os.path.dirname(package_root)
 kibana_dir = os.path.join(package_root, 'specs', 'kibana-7.8.1')
